requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.15.0
Pillow>=6.2.2
pyparsing==2.4.2
python-dateutil==2.8.1
pytz==2018.4
six==1.12.0

torchsummary==1.5.1
